requirements.txt
Flask
openai
python-dotenv
gunicorn
google-generativeai
